ink-detection: organize code into data, model, training, and inference packages
