+++ vesuvius/src/vesuvius/ink_detection/data/geometry.py
