+++ vesuvius/src/vesuvius/ink_detection/data/normalization.py
