+++ vesuvius/src/vesuvius/ink_detection/data/patch_cache.py
