+++ vesuvius/src/vesuvius/ink_detection/data/patch_finding_default.py
