+++ vesuvius/src/vesuvius/ink_detection/data/segment.py
