+++ vesuvius/src/vesuvius/ink_detection/models/checkpoint.py
