+++ vesuvius/src/vesuvius/ink_detection/models/hybrid_3d2d.py
