+++ vesuvius/src/vesuvius/ink_detection/models/input_padding.py
