+++ vesuvius/src/vesuvius/ink_detection/training/deep_supervision.py
