+++ vesuvius/src/vesuvius/ink_detection/training/stitching.py
