+++ vesuvius/src/vesuvius/ink_detection/training/visualization.py
